[JBIDE-23499] fixed and merged ConnectionWizardLinkTest
* fixed ConnectionWizardLinkTest to login into CDK via OAuth
  (was trying to use openshift 3 online preview with a disabled account)
* renamed the single test method in it and moved into
CreateNewConnectionTest
* extracted labels to OpenShiftLabel
