+++ tests/org.jboss.tools.openshift.ui.bot.test/src/org/jboss/tools/openshift/ui/bot/test/OpenShift3BotTests.java
@@ -38,7 +38,6 @@ import org.jboss.tools.openshift.ui.bot.test.application.v3.create.DeploymentTes
 import org.jboss.tools.openshift.ui.bot.test.common.OCBinaryLocationTest;
 import org.jboss.tools.openshift.ui.bot.test.connection.v3.ConnectionPropertiesTest;
 import org.jboss.tools.openshift.ui.bot.test.connection.v3.ConnectionWizardHandlingTest;
-import org.jboss.tools.openshift.ui.bot.test.connection.v3.ConnectionWizardLinksTest;
 import org.jboss.tools.openshift.ui.bot.test.connection.v3.CreateNewConnectionTest;
 import org.jboss.tools.openshift.ui.bot.test.connection.v3.OpenNewConnectionShellTest;
 import org.jboss.tools.openshift.ui.bot.test.connection.v3.RemoveConnectionTest;
@@ -67,7 +66,6 @@ import org.junit.runners.Suite.SuiteClasses;
 	
 	// Connection
 	OpenNewConnectionShellTest.class,
-	ConnectionWizardLinksTest.class,
 	CreateNewConnectionTest.class,
 	RemoveConnectionTest.class,
 	ConnectionWizardHandlingTest.class,
