+++ tests/org.jboss.tools.openshift.ui.bot.test/src/org/jboss/tools/openshift/ui/bot/test/OpenShift3SmokeBotTests.java
@@ -37,7 +37,6 @@ import org.jboss.tools.openshift.ui.bot.test.connection.v3.ConnectionPropertiesT
 import org.jboss.tools.openshift.ui.bot.test.connection.v3.ConnectionWizardHandlingTest;
 import org.jboss.tools.openshift.ui.bot.test.connection.v3.CreateNewConnectionTest;
 import org.jboss.tools.openshift.ui.bot.test.connection.v3.OpenNewConnectionShellTest;
-import org.jboss.tools.openshift.ui.bot.test.connection.v3.ConnectionWizardLinksTest;
 import org.jboss.tools.openshift.ui.bot.test.connection.v3.RemoveConnectionTest;
 import org.jboss.tools.openshift.ui.bot.test.connection.v3.ShowConnectionInWebConsoleTest;
 import org.jboss.tools.openshift.ui.bot.test.connection.v3.StoreConnectionTest;
@@ -64,7 +63,6 @@ import org.junit.runners.Suite.SuiteClasses;
 	
 	// Connection
 	OpenNewConnectionShellTest.class,
-	ConnectionWizardLinksTest.class,
 	CreateNewConnectionTest.class,
 	RemoveConnectionTest.class,
 	ConnectionWizardHandlingTest.class,
